refactor(rpc): move auth errors into module
