+++ crates/rpc/src/auth/mod.rs
@@ -2,7 +2,7 @@
 
 mod token;
 
+pub use crate::error::AuthError;
 pub use token::{
-    AuthContext, AuthError, AuthorizationToken, X_AUTHORIZATION_TOKEN, build_token_fields,
-    parse_auth_header,
+    AuthContext, AuthorizationToken, X_AUTHORIZATION_TOKEN, build_token_fields, parse_auth_header,
 };
