+++ crates/rpc/src/lib.rs
@@ -8,6 +8,7 @@
 
 pub mod auth;
 pub mod config;
+pub mod error;
 pub mod filter;
 pub mod handlers;
 pub mod policy;
